Create projects/index and /show views, plus related
Main update here was the creation of /index and /show views for
projects. At the same time, though, instance variables needed to
be updated in the projects controller, and 'normalize.css'
needed to be added to the assets initializer. Small changes were
also made to the _header and _nav partials.